clarify and simplify api token tests
